requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0

# Optional - for AI features
openai==1.10.0
google-generativeai==0.3.2

# For better async performance
aiohttp==3.9.1
requests==2.31.0
